+++ djangocms_moderation/signals.py
@@ -6,3 +6,5 @@ confirmation_form_submission = django.dispatch.Signal()
 submitted_for_review = django.dispatch.Signal()
 
 published = django.dispatch.Signal()
+
+unpublished = django.dispatch.Signal()